Code Review II und III: Language Server 604 auf 3, Werkzeugkasten 1.263 auf 1.079
Doppelcode 66 auf 27: Vorwärtskinematik des Rigs aus vier Kopien in
ModelPhysik/knochenwelt.py (Gegenprobe gegen die alten Fassungen am DEF-Rig
bitgleich); Videoschreiber.rendern/schreiben und Bakedatei.auf_dem_boden für
die vier Filmskripte; Hautverdeckung erbt von Figurhaut, Umaanimation von
Eigenanimation; Modell.skelettBauen, Netzantwort.hautgewichte,
Charakterdaten.zustand (die WebSocket-Kanäle luden je Verbindung eigene
MorphData); Kontextmenue im Animationsmenü; Buehnenschatten (Theatre);
Modellkindspur für Mimik- und Script-Spur; Studio-Untermenüs Länge und
Animation als Includes mit _studiovorlage für die Textprüfungen; Grundfall
_effektseite für die Effekte-Seitentests; pruefe() als Vorspann in Jsmodul.

Weiter: szenarien 19 auf 1 (18 Testnamen), jsfunktionen, jsregistrierung,
getattr-namen und leserzahl auf 0, klassenplan 3 auf 1, rueckgabedict 3 auf 1
(Linse und Garmentanfrage als namedtuple), Sortierschlüssel der Hilfetabellen
als ganze Zahl, pyright-Vermerke für Pakete anderer Umgebungen, Sicher.wert
für Optionales in Prüfungen.

Suite 2.467 grün, Konformität 118, angefasste LongRunner 57 grün.

+++ core/tests/longrunner/test_mhfigur.py
@@ -69,7 +69,7 @@ class MhfigurTest(SimpleTestCase):
         u"""Positives Volumen = Normalen nach aussen (siehe Mhbasisnetz)."""
         daten = self._netz()
         punkte = MhfigurTest._feld(daten['vertices'], np.float32, 3).astype(np.float64)
-        dreiecke = MhfigurTest._feld(daten['faces'], np.uint32, 3)
+        dreiecke = MhfigurTest._feld(daten['faces'], np.uint32, 3).astype(np.int64)
         v0, v1, v2 = (punkte[dreiecke[:, i]] for i in range(3))
         volumen = float(np.einsum('ij,ij->i', v0, np.cross(v1, v2)).sum() / 6.0)
         self.assertGreater(volumen, 0.03, 'Umlaufsinn gedreht')
